+++ src/cryptoadvance/specter/helpers.py
@@ -1,9 +1,10 @@
+# Additions to helpers.py are deprecated. Please place your helpers in utils/somewhere
+
 import binascii
 import collections
 import copy
 import hashlib
 import hmac
-import json
 import logging
 import os
 import six
@@ -21,7 +22,6 @@ from .util.bcur import bcur_decode
 import threading
 from io import BytesIO
 import re
-from datetime import datetime
 
 logger = logging.getLogger(__name__)
 
@@ -308,17 +308,3 @@ def get_address_from_dict(data_dict):
     if addr and addr != "Fee":
         return addr
     raise RuntimeError(f"Missing address info in object {data_dict}")
-
-
-def robust_json_dumps(obj):
-    def default(o):
-        if isinstance(o, datetime):
-            return o.timestamp()
-        if isinstance(o, set):
-            return list(o)
-        logger.warning(
-            f"robust_json_dumps could not convert {o} of type {type(o)}.  Converting to string instead."
-        )
-        return str(o)
-
-    return json.dumps(obj, default=default)
