+++ src/cryptoadvance/specter/server_endpoints/wallets_api.py
@@ -32,7 +32,7 @@ from ..util.price_providers import get_price_at
 from ..util.tx import decoderawtransaction
 from embit.descriptor.checksum import add_checksum
 from ..global_search import do_global_search
-from ..helpers import robust_json_dumps
+from ..util.common import robust_json_dumps
 
 logger = logging.getLogger(__name__)
